+++ CHANGELOG.md
@@ -9,9 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Reports for timed-out tests and hooks.
+- Reports for timed-out tests and hooks enabled with `WithTimeout` option.
 - Option `WithGroupHooks` to configure `BeforeAll` & `AfterAll` hooks groupping in reports.
-- Option `WithTimeouts` to disable timeout handling.
 
 ## [1.2.2] - 2026-06-18